refactor: remove Pipeline orchestrator, replace with standalone pollers (#311) (#342)
Remove the scheduler Pipeline, tiers (1-3), WatchQueue, and bridge in
favour of standalone goroutines wired directly in main.go. The Pipeline
was the last remnant of the pre-NATS architecture — Tier 3 was already
replaced by the NATS KV state worker, and the WatchQueue by NATS KV.

What changed:
- Delete pipeline.go, tier1.go, tier2.go, tier3.go, queue.go and all
  their tests (bridge_test, tier1_test, tier2_test, tier3_test, queue_test)
- Add types.go to keep exported types (Tier2PR, WatchItem, ItemSnapshot,
  interfaces) that external consumers still reference
- Replace buildPipeline/pipe.Start/pipe.Stop in main.go with
  startPollers/pollerCancel/pollerWg using standalone goroutines
- Inline discovery publishing (sendDiscoveryRepos), NATS bridge
  (bridgeDiscovery), and PR/issue polling (runTier2) as package-level
  functions in main.go
- NATS workers now reference the standalone limiter directly instead of
  pipe.Limiter()
- Config reload stops/starts pollers via context cancellation

Kept: ratelimit.go, scheduler.go, SSE broker — still used by dependents.
All existing tests pass.
